Adding support for SOLID donut graphs
This single-line fix aims to add tooltip support for donut graphs created with the 'donutSolid:true' parameter.

Without this patch, the tooltip plugin only looks to SVG DOM elements having the '.ct-slice-donut' class, then leading the tooltip plugin to be not not functional on SLoid Donut charts.

As a chart created with the **'donutSolid:true'** parameter lead Chartist to assign the **'.ct-slice-donut-solid'** class to the SVG DOM elements slices, it is therefore needed to perform an additional test in the tooltip plugin.

This patch will test if `chart.options.donutSolid` is `true`, and then define the `tooltipSelector` to **`'.ct-slice-donut-solid'`** instead of `'.ct-slice-donut'`, which is reliable only when `chart.options.donutSolid` is `false`.

(cherry picked from commit dec622fcdd78b542c6efdb4d4b9588bf54c58434)